+++ blog-core/src/main/java/com/zyd/blog/business/service/impl/BizTypeServiceImpl.java
@@ -5,7 +5,10 @@ import com.github.pagehelper.PageInfo;
 import com.zyd.blog.business.entity.Type;
 import com.zyd.blog.business.service.BizTypeService;
 import com.zyd.blog.business.vo.TypeConditionVO;
+import com.zyd.blog.framework.exception.ZhydException;
+import com.zyd.blog.persistence.beans.BizArticle;
 import com.zyd.blog.persistence.beans.BizType;
+import com.zyd.blog.persistence.mapper.BizArticleMapper;
 import com.zyd.blog.persistence.mapper.BizTypeMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -31,6 +34,8 @@ public class BizTypeServiceImpl implements BizTypeService {
 
     @Autowired
     private BizTypeMapper bizTypeMapper;
+    @Autowired
+    private BizArticleMapper bizArticleMapper;
 
     @Override
     public PageInfo<Type> findPageBreakByCondition(TypeConditionVO vo) {
@@ -83,6 +88,13 @@ public class BizTypeServiceImpl implements BizTypeService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public boolean removeByPrimaryKey(Long primaryKey) {
+
+        BizArticle article = new BizArticle();
+        article.setTypeId(primaryKey);
+        List<BizArticle> articles = bizArticleMapper.select(article);
+        if (!CollectionUtils.isEmpty(articles)) {
+            throw new ZhydException("当前分类下存在文章信息，禁止删除！");
+        }
         return bizTypeMapper.deleteByPrimaryKey(primaryKey) > 0;
     }
